backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.10
pydantic-settings==2.10.1
python-multipart==0.0.20
pypdf==6.0.0
